divide Oj::Formatter from Rack::CamelSnake
