+++ spec/rack/camel_snake_spec.rb
@@ -67,40 +67,4 @@ describe Rack::CamelSnake do
     end
   end
 
-  describe 'to_snake' do
-    it 'convert camel case into snake case' do
-      app.send(:to_snake, 'CamelCase').should eq 'camel_case'
-      app.send(:to_snake, 'CAMELCase').should eq 'camel_case'
-    end
-  end
-
-  describe 'to_camel' do
-    it 'convert snake case into camel case' do
-      %w(_snake_case snake_case snake___case snake_case_).each do |word|
-        app.send(:to_camel, word).should eq 'snakeCase'
-      end
-    end
-  end
-
-  describe 'formatter' do
-    let!(:snake_hash){ { 'is_done' => 'hoge', 'order' => 1, 'task_title' => 'title' } }
-    let!(:camel_hash){ { 'isDone'  => 'hoge', 'order' => 1, 'taskTitle'  => 'title' } }
-    let(:snake_array){ [ snake_hash, snake_hash, snake_hash ] }
-    let(:camel_array){ [ camel_hash, camel_hash, camel_hash ] }
-
-    context 'given :to_camel' do
-      it 'converts keys into camelCase, and the keys should be a string.' do
-        app.send(:formatter, snake_hash,  app.send(:key_converter, :to_camel)).should eq camel_hash
-        app.send(:formatter, snake_array, app.send(:key_converter, :to_camel)).should eq camel_array
-      end
-    end
-
-    context 'given :to_snake' do
-      it 'converts keys into snake_case, and the keys should be a symbol.' do
-        app.send(:formatter, camel_hash,  app.send(:key_converter, :to_snake)).should eq snake_hash
-        app.send(:formatter, camel_array, app.send(:key_converter, :to_snake)).should eq snake_array
-      end
-    end
-  end
-
 end
